Move croc to ci folder.
